make install faster

+++ install-setup.js
@@ -135,56 +135,120 @@ let binLocation = "/usr/local/bin/";
 
 let search = "/lib/systemd/system/";
 let replacer = new RegExp(search, "g");
+let systemdUnitDir = "/lib/systemd/system/";
 
 let existingServices = String(execSync(`find /lib/systemd/system -name 'smartsoundsync*'`));
 existingServices = existingServices.replace(replacer, "");
 existingServices = existingServices.split(/\r?\n/);
 
-existingServices.forEach(function (value, index) {
-    if (value.length > 0) {
-        try {
-            execSyncPrint(`systemctl stop ${value}`);
-        } catch (error) {
-            console.log("Error: could not stop", value, error);
-        }
+function getServiceUnitPath(serviceName) {
+    return `${systemdUnitDir}${serviceName}`;
+}
 
-        try {
-            execSyncPrint(`systemctl disable ${value}`);
-        } catch (error) {
-            console.log("Error: could not disable", value, error);
-        }
+function removeServiceUnitFile(serviceName) {
+    let serviceUnitPath = getServiceUnitPath(serviceName);
 
-        try {
-            execSyncPrint(`rm /lib/systemd/system/${value}`);
-        } catch (error) {
-            console.log("Error: could not remove", value, error);
-        }
+    if (!fs.existsSync(serviceUnitPath)) {
+        return;
     }
-});
 
-function writeServiceFile(serviceName, serviceTemplate) {
-    console.log(`writing service file ${serviceName}`);
-    fs.writeFileSync(`${installLocation}/${serviceName}`, serviceTemplate, "utf8");
+    execSyncPrint(`rm ${serviceUnitPath}`);
+}
+
+function stopServiceIfActive(serviceName) {
+    if (!serviceIsActive(serviceName)) {
+        return;
+    }
 
     try {
-        execSyncPrint(`mv ${installLocation}/${serviceName} /lib/systemd/system/${serviceName}`);
+        execSyncPrint(`systemctl stop ${serviceName}`);
     } catch (error) {
-        console.log(`Error: error moving service file to systemd. deleting template ${serviceName}`, error);
-        execSyncPrint(`rm ${installLocation}/${serviceName}`);
+        console.log("Error: could not stop", serviceName, error);
     }
 }
 
-function serviceStart(serviceName) {
+function serviceIsEnabled(serviceName) {
+    try {
+        execSync(`systemctl is-enabled --quiet ${serviceName}`);
+        return true;
+    } catch (error) {
+        return false;
+    }
+}
+
+function disableServiceIfEnabled(serviceName) {
+    if (!serviceIsEnabled(serviceName)) {
+        return;
+    }
+
+    try {
+        execSyncPrint(`systemctl disable ${serviceName}`);
+    } catch (error) {
+        console.log("Error: could not disable", serviceName, error);
+    }
+}
+
+function unmaskServiceIfExists(serviceName) {
+    if (!serviceUnitExists(serviceName)) {
+        console.log(`service ${serviceName} does not exist, skipping unmask`);
+        return;
+    }
+
+    try {
+        execSyncPrint(`systemctl unmask ${serviceName}`);
+    } catch (error) {
+        console.log(`Error: could not unmask ${serviceName}`, error);
+    }
+}
+
+function enableServiceIfNeeded(serviceName) {
+    if (serviceIsEnabled(serviceName)) {
+        return;
+    }
+
     try {
         execSyncPrint(`systemctl enable ${serviceName}`);
     } catch (error) {
         console.log("Error: could not enable", serviceName, error);
     }
+}
 
-    try {
-        execSyncPrint(`systemctl start ${serviceName}`);
-    } catch (error) {
-        console.log("Error: could not start", serviceName, error);
+function writeServiceFileIfChanged(serviceName, serviceTemplate) {
+    let serviceUnitPath = getServiceUnitPath(serviceName);
+    let existingTemplate = "";
+
+    if (fs.existsSync(serviceUnitPath)) {
+        existingTemplate = fs.readFileSync(serviceUnitPath, "utf8");
+    }
+
+    if (existingTemplate === serviceTemplate) {
+        console.log(`service file unchanged ${serviceName}`);
+        return false;
+    }
+
+    console.log(`writing service file ${serviceName}`);
+    fs.writeFileSync(serviceUnitPath, serviceTemplate, "utf8");
+    return true;
+}
+
+function applyServiceState(serviceName, unitChanged) {
+    enableServiceIfNeeded(serviceName);
+
+    if (unitChanged && serviceIsActive(serviceName)) {
+        try {
+            execSyncPrint(`systemctl restart ${serviceName}`);
+        } catch (error) {
+            console.log("Error: could not restart", serviceName, error);
+        }
+        return;
+    }
+
+    if (!serviceIsActive(serviceName)) {
+        try {
+            execSyncPrint(`systemctl start ${serviceName}`);
+        } catch (error) {
+            console.log("Error: could not start", serviceName, error);
+        }
     }
 }
 
@@ -211,6 +275,17 @@ function makeEcasoundConfig() {
     execSync(`cp ${installLocation}/config_examples/${ecasoundChainSetupFileName} ${ecasoundChainSetupFilePath}`);
     execSync(`cp ${installLocation}/config_examples/${ecasoundFilterFileName} ${ecasoundFilterFilePath}`);
     execSyncPrint(`chown -R ${installLocationUser} ${configFileDir}`);
+}
+
+function stopOnlyCleanup() {
+    existingServices.forEach(function (serviceName) {
+        if (serviceName.length === 0) {
+            return;
+        }
+
+        stopServiceIfActive(serviceName);
+        disableServiceIfEnabled(serviceName);
+    });
 }
 
 function ensureRustToolchain() {
@@ -384,6 +459,8 @@ function getNtpStatusServiceSettings(settings) {
 }
 
 function standardizeOnNtpsec() {
+    let removedConflictingNtpPackages = false;
+
     if (!packageIsInstalled("ntpsec")) {
         execSyncPrint(`apt install ntpsec -y`);
     } else {
@@ -397,6 +474,7 @@ function standardizeOnNtpsec() {
     if (packageIsInstalled("ntp")) {
         try {
             execSyncPrint(`apt purge ntp -y`);
+            removedConflictingNtpPackages = true;
         } catch (error) {
             console.log("Error: could not purge ntp", error);
         }
@@ -405,18 +483,21 @@ function standardizeOnNtpsec() {
     if (packageIsInstalled("chrony")) {
         try {
             execSyncPrint(`apt purge chrony -y`);
+            removedConflictingNtpPackages = true;
         } catch (error) {
             console.log("Error: could not purge chrony", error);
         }
     }
 
-    try {
-        execSyncPrint(`apt autoremove -y`);
-    } catch (error) {
-        console.log("Error: could not run apt autoremove", error);
+    if (removedConflictingNtpPackages) {
+        try {
+            execSyncPrint(`apt autoremove -y`);
+        } catch (error) {
+            console.log("Error: could not run apt autoremove", error);
+        }
+    } else {
+        console.log("no conflicting ntp packages removed. Skipping apt autoremove.");
     }
-
-    unmaskEnableServiceIfExists("ntpsec");
 }
 
 const librespotRepoZip = "https://github.com/mikeszila/librespot/archive/dev.zip";
@@ -444,7 +525,9 @@ function setInstalledLibrespotCommit(commit) {
     fs.writeFileSync(librespotCommitFile, `${commit}\n`, "utf8");
 }
 
-if (!stopOnly) {
+if (stopOnly) {
+    stopOnlyCleanup();
+} else {
     if (fs.existsSync("/usr/local/etc/smartsoundsyncconf.js")) {
         console.log("found old style config.  Converting to new style at", configFilePath);
         execSync(`mkdir -p /usr/local/etc/smartsoundsync/`);
@@ -584,7 +667,8 @@ if (!stopOnly) {
         console.log("no changes to ntp config.");
     }
 
-    unmaskEnableServiceIfExists(ntpConfigDetails.serviceName);
+    unmaskServiceIfExists(ntpConfigDetails.serviceName);
+    enableServiceIfNeeded(ntpConfigDetails.serviceName);
     if (ntpConfigChanged) {
         execSyncPrint(`systemctl restart ${ntpConfigDetails.serviceName}`);
     } else if (!serviceIsActive(ntpConfigDetails.serviceName)) {
@@ -772,7 +856,9 @@ if (!stopOnly) {
 
     let serviceName = "";
     let serviceTemplate = "";
-    let servicesToStart = [];
+    let desiredServiceFiles = [];
+    let servicesToActivate = [];
+    let changedServices = [];
     let execArguments = "";
     let priority = 1;
 
@@ -800,8 +886,11 @@ WantedBy=multi-user.target
 `;
     serviceName = `smartsoundsynccommon.service`;
 
-    writeServiceFile(serviceName, serviceTemplate);
-    servicesToStart.push(serviceName);
+    if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+        changedServices.push(serviceName);
+    }
+    desiredServiceFiles.push(serviceName);
+    servicesToActivate.push(serviceName);
 
     execArguments = `"${execArgumentsParse(ntpStatusSettings)}"`;
 
@@ -824,8 +913,11 @@ WantedBy=multi-user.target
 `;
     serviceName = `smartsoundsyncntp.service`;
 
-    writeServiceFile(serviceName, serviceTemplate);
-    servicesToStart.push(serviceName);
+    if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+        changedServices.push(serviceName);
+    }
+    desiredServiceFiles.push(serviceName);
+    servicesToActivate.push(serviceName);
 
     if (settings.controller) {
         execArguments = "";
@@ -856,8 +948,11 @@ WantedBy=multi-user.target
 `;
         serviceName = `smartsoundsynccontrol.service`;
 
-        writeServiceFile(serviceName, serviceTemplate);
-        servicesToStart.push(serviceName);
+        if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+            changedServices.push(serviceName);
+        }
+        desiredServiceFiles.push(serviceName);
+        servicesToActivate.push(serviceName);
     }
 
     if (settings.sink) {
@@ -890,10 +985,14 @@ WantedBy=multi-user.target
 `;
         serviceName = `smartsoundsyncsink.service`;
 
-        writeServiceFile(serviceName, serviceTemplate);
+        if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+            changedServices.push(serviceName);
+        }
+
+        desiredServiceFiles.push(serviceName);
 
         if (!hasHifiberryDacDSP) {
-            servicesToStart.push(serviceName);
+            servicesToActivate.push(serviceName);
         }
     }
 
@@ -932,8 +1031,11 @@ WantedBy=multi-user.target
 `;
                 serviceName = `smartsoundsyncspdif${sourceSettings.audioSourceDisplayName}.service`;
 
-                writeServiceFile(serviceName, serviceTemplate);
-                servicesToStart.push(serviceName);
+                if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+                    changedServices.push(serviceName);
+                }
+                desiredServiceFiles.push(serviceName);
+                servicesToActivate.push(serviceName);
             }
 
             if (sourceSettings.audioSourceType === "Spotify") {
@@ -945,7 +1047,6 @@ Requires=network-online.target avahi-daemon.service
 [Service]
 Type=simple
 WorkingDirectory=${installLocation}
-ExecStartPre=/bin/sleep 5
 ExecStart=/usr/bin/node ${installLocation}/pipelibrespottoudp.js "${execArgumentsParse(sourceSettings)}" 
 TimeoutStopSec=5
 
@@ -957,8 +1058,11 @@ WantedBy=multi-user.target
 `;
                 serviceName = `smartsoundsyncspotify${sourceSettings.audioSourceDisplayName}.service`;
 
-                writeServiceFile(serviceName, serviceTemplate);
-                servicesToStart.push(serviceName);
+                if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+                    changedServices.push(serviceName);
+                }
+                desiredServiceFiles.push(serviceName);
+                servicesToActivate.push(serviceName);
             }
 
             if (sourceSettings.audioSourceType === "Airplay") {
@@ -985,18 +1089,45 @@ WantedBy=multi-user.target
 `;
                 serviceName = `smartsoundsyncairplay${sourceSettings.audioSourceDisplayName}.service`;
 
-                writeServiceFile(serviceName, serviceTemplate);
-                servicesToStart.push(serviceName);
+                if (writeServiceFileIfChanged(serviceName, serviceTemplate)) {
+                    changedServices.push(serviceName);
+                }
+                desiredServiceFiles.push(serviceName);
+                servicesToActivate.push(serviceName);
             }
         });
     }
 
-    execSyncPrint(`systemctl daemon-reload`);
+    let removedServices = [];
 
-    servicesToStart.forEach(function (value, index) {
-        serviceStart(value);
+    existingServices.forEach(function (serviceName) {
+        if (serviceName.length === 0) {
+            return;
+        }
+
+        if (desiredServiceFiles.includes(serviceName)) {
+            return;
+        }
+
+        console.log(`removing obsolete service ${serviceName}`);
+        stopServiceIfActive(serviceName);
+        disableServiceIfEnabled(serviceName);
+        removeServiceUnitFile(serviceName);
+        removedServices.push(serviceName);
+    });
+
+    if (changedServices.length > 0 || removedServices.length > 0) {
+        execSyncPrint(`systemctl daemon-reload`);
+    } else {
+        console.log("no service unit changes detected. Skipping systemctl daemon-reload.");
+    }
+
+    servicesToActivate.forEach(function (serviceName) {
+        applyServiceState(serviceName, changedServices.includes(serviceName));
     });
 }
 
-execSyncPrint(`chown -R ${installLocationUser} ${installLocation}`);
-execSyncPrint(`chown -R ${installLocationUser} ${configFileDir}`);
+if (!stopOnly) {
+    execSyncPrint(`chown -R ${installLocationUser} ${installLocation}`);
+    execSyncPrint(`chown -R ${installLocationUser} ${configFileDir}`);
+}
